% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlinbears.R
\docType{data}
\name{berlinbears}
\alias{berlinbears}
\title{Bears bears dataframe}
\format{
\subsection{\code{berlinbears}}{

A data frame with 1000 rows and 8 columns:
\describe{
\item{species}{name of species}
\item{gender}{gender of bear}
\item{weight}{weight of bear in kg}
\item{favorite_food, favorite_fruit}{survey questions}
\item{likert1, likert2, likert3}{example of likert questions}
...
}
}
}
\usage{
berlinbears
}
\description{
A "survey" of bears in Berlin
Report ...
}
\keyword{datasets}
