% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional functions.R,
%   R/conditional_functions.R
\name{conditional_median}
\alias{conditional_median}
\title{Calculate conditional medians}
\usage{
conditional_median(data, median_var, category_var)

conditional_median(data, median_var, category_var)
}
\arguments{
\item{data}{Dataframe that contains both \code{median_var} and \code{category_var}}

\item{median_var}{data that will be used in the calculation of medians}

\item{category_var}{The categories across which the function will compute medians}
}
\value{
a  dataframe with the means of \code{median_var} for each level of \code{category_var}

a  dataframe with the means of \code{median_var} for each level of \code{category_var}
}
\description{
Calculate conditional medians

Calculate conditional medians
}
