% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_styling.R
\docType{data}
\name{dezim_colors}
\alias{dezim_colors}
\title{Vector of colors for DeZIM Plots}
\format{
\subsection{\code{dezim_colors}}{

A data frame with 500 rows and 22 columns describing bears and thier prefrences:
\describe{
\item{species}{name of species}
...
}
}
}
\usage{
dezim_colors
}
\description{
Vector of colors for DeZIM Plots
}
\keyword{datasets}
