% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_styling.R
\name{dezim_style}
\alias{dezim_style}
\title{DeZIM Graphics Style}
\usage{
dezim_style(
  calibri = FALSE,
  title_size = 16,
  sub_title_size = 14,
  x_axis_title_size = 14,
  y_axis_title_size = 14,
  x_axis_text_size = 12,
  y_axis_text_size = 12,
  legend_text_size = 12
)
}
\arguments{
\item{calibri}{Use calibri as font family}

\item{title_size}{Size of title}

\item{sub_title_size}{Size of subtitle}

\item{x_axis_title_size}{size of x axis title}

\item{y_axis_title_size}{size of y axis title}

\item{x_axis_text_size}{size of x axis text}

\item{y_axis_text_size}{size of y axis text}

\item{legend_text_size}{Size of text in legend}
}
\value{
No return, just a theme to add to ggplotfiles
}
\description{
DeZIM Graphics Style
}
