% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_of_answers.R
\name{generate_unique_level_list}
\alias{generate_unique_level_list}
\title{Generate unique level list}
\usage{
generate_unique_level_list(dataframe, column_names)
}
\arguments{
\item{dataframe}{.}

\item{column_names}{.}
}
\value{
.
}
\description{
Generate unique level list
}
