% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert.R
\name{likert_graph}
\alias{likert_graph}
\title{Quickly plot likert data (3,5,7,9 levels)}
\usage{
likert_graph(
  dataset,
  question,
  na.rm = TRUE,
  labels = NULL,
  colors = NULL,
  weights = NULL
)
}
\arguments{
\item{dataset}{Dataframe or a tibble containing the \code{question} column to be analyzed}

\item{question}{The columns that contain each of the response options, can be selected by
using \strong{tidyselect} semanatics or providing a vector of column names or numbers}

\item{weights}{Optional column containing survey weights}
}
\value{
example return
}
\description{
Quickly plot likert data (3,5,7,9 levels)
}
