% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/techdebt.R
\name{likert_graph_old}
\alias{likert_graph_old}
\title{Quickly plot likert data (3,5,7,9 levels)}
\usage{
likert_graph_old(
  data,
  labels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agre"),
  colors = c("#d73027", "#E36A64", "#FEEBD7", "#8ECF8C", "#66bd63")
)
}
\arguments{
\item{data}{Summerized likert data (first column must be names)}

\item{colors}{example text}

\item{levels}{example text}
}
\value{
example return
}
\description{
Quickly plot likert data (3,5,7,9 levels)
}
