% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_analysis.R
\name{likert_summary}
\alias{likert_summary}
\title{Likert Summary
Summarize likert data}
\usage{
likert_summary(data, high_to_low = TRUE, order_rows = FALSE)
}
\arguments{
\item{data}{either a dataframe or a vector of likert data}

\item{high_to_low}{If true, the highest numbers will appear in the first column}

\item{order_rows}{If true rows will be ordered by largest to smallest}
}
\value{
A dataframe of percentages corresponding to the share of each category
}
\description{
Likert Summary
Summarize likert data
}
