% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix_mean}
\alias{matrix_mean}
\title{Matrix Mean Plot}
\usage{
matrix_mean(
  dataset,
  question,
  group_by = NULL,
  subgroups_to_exclude = NULL,
  weights = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{dataset}{The input dataframe (or tibble) of survey questions}

\item{question}{The columns that contain each of the response options for a
question, can be selected by using \strong{tidyselect} semanatics or providing a
vector of column names or numbers}

\item{group_by}{Optional variable to group the analysis. If provided, the
frequencies and counts will be calculated within each subgroup.}

\item{subgroups_to_exclude}{Optional vector specifying subgroups to exclude
from the analysis.}

\item{weights}{Optional variable containing survey weights. If provided,
frequencies and counts will be weighted accordingly.}

\item{na.rm}{Logical indicating whether to remove NA values from \code{question}
before analysis.}
}
\value{
A ggplot object representing the mean plot.
}
\description{
Visualizes the mean and standard deviation of responses to a specified question.
If grouping is specified, the plot compares the means across different groups.
}
\seealso{
Other matrix questions: 
\code{\link{matrix_freq}()},
\code{\link{matrix_likert}()},
\code{\link{matrix_table}()}
}
\concept{matrix questions}
