% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix_table}
\alias{matrix_table}
\title{Tabulate Array/Matrix questions}
\usage{
matrix_table(
  dataset,
  question,
  group_by = NULL,
  subgroups_to_exclude = NULL,
  weights = NULL
)
}
\arguments{
\item{dataset}{Dataframe or a tibble containing the \code{question} column to be analyzed}

\item{question}{The columns that contain each of the response options, can be selected by
using \strong{tidyselect} semanlatics or providing a vector of column names or numbers}

\item{group_by}{Name of column in \code{dataset} used to partition the analysis into subgroups}

\item{subgroups_to_exclude}{vector that contains level(s) of \code{group_by} variable to exclude}

\item{weights}{Optional column containing survey weights}
}
\value{
placeholder
}
\description{
Tabulate Array/Matrix questions
}
