% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplechoice.R
\name{multichoice_graph}
\alias{multichoice_graph}
\title{Title}
\usage{
multichoice_graph(
  dataset,
  question,
  group_by = NULL,
  subgroups_to_exclude = NULL,
  weights = NULL
)
}
\arguments{
\item{dataset}{Dataframe or a tibble containing the \code{question} column to be analyzed}

\item{question}{Name of the column in the \code{dataset} to be summerized}

\item{group_by}{Name of column in \code{dataset} used to partition the analysis into subgroups}

\item{subgroups_to_exclude}{vector that contains level(s) of \code{group_by} variable to exclude}

\item{weights}{Optional column containing survey weights}
}
\value{
placehodler
}
\description{
Title
}
