% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlechoice.R
\name{singlechoice_summary}
\alias{singlechoice_summary}
\title{Summarize count and frequency of a single choice survey question}
\usage{
singlechoice_summary(
  dataset,
  question,
  subgroup = NULL,
  levels_to_exclude = NULL,
  weights = NULL
)
}
\arguments{
\item{dataset}{Dataframe or a tibble containing the \code{question} to be analyzed}

\item{question}{Name of the column in the \code{dataset} to be summerized}

\item{subgroup}{Optional variable to stratify the frequencies of \code{question} variable}

\item{levels_to_exclude}{Optional vector of levels of \code{subgroup} to exclude}

\item{weights}{Optional column containing survey weights}
}
\value{
Dataframe of count and frequency for each
}
\description{
Summarize count and frequency of a single choice survey question
}
