% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_of_answers.R
\name{struct_level}
\alias{struct_level}
\title{Get structure of answer levels}
\usage{
struct_level(dataframe)
}
\arguments{
\item{dataframe}{.}
}
\value{
.
}
\description{
Get structure of answer levels
}
