% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_of_answers.R
\name{structure_answers}
\alias{structure_answers}
\title{Structure of answers}
\usage{
structure_answers(dataframe, category_threshold = 20)
}
\arguments{
\item{dataframe}{placeholder}

\item{category_threshold}{The number of unique answers to a question above which it is considered unstructured}
}
\value{
a dataframe indicating how many questions are an integer, numeric, logical, unstructured, or categorical
}
\description{
Structure of answers
}
