% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_of_answers.R
\name{structure_categories}
\alias{structure_categories}
\title{Structure of categorical answeres}
\usage{
structure_categories(dataframe, category_threshold = 20)
}
\arguments{
\item{dataframe}{placeholder}

\item{category_threshold}{The number of unique answers to a question above which it is considered unstructured}
}
\value{
a list grouped by how many unique answers each variable has
}
\description{
Structure of categorical answeres
}
