% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_and_replace.R
\name{Zfind}
\alias{Zfind}
\title{Find answer columns that match given answer choices}
\usage{
Zfind(dataframe, answer_pattern, categorical_only = TRUE, column_names = FALSE)
}
\arguments{
\item{dataframe}{placeholder}

\item{answer_pattern}{answers to look-up}

\item{categorical_only}{reccomended. only look at variables considered to be categorical}

\item{column_names}{placeholder}
}
\value{
column numbers that match the given string, or if column_names = TRUE then the names of the columns
}
\description{
Find answer columns that match given answer choices
}
