% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_and_replace.R
\name{Zreplace}
\alias{Zreplace}
\title{Zreplace
Replace a given string with a new string in selected columns}
\usage{
Zreplace(dataframe, columns_to_replace, old_text, new_text)
}
\arguments{
\item{dataframe}{placeholder}

\item{columns_to_replace}{placeholder}

\item{old_text}{placeholder}

\item{new_text}{placeholder}
}
\value{
dataset with values replaced in selected columns
}
\description{
Zreplace
Replace a given string with a new string in selected columns
}
