% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_of_answers.R
\name{Zstructure}
\alias{Zstructure}
\title{Zstructure}
\usage{
Zstructure(dataframe, mode, category_threshold = 20)
}
\arguments{
\item{dataframe}{later}

\item{mode}{later}

\item{category_threshold}{later}
}
\value{
later
}
\description{
Zstructure
}
