% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figures.R, R/graphics_styling.R
\name{dezim_style}
\alias{dezim_style}
\title{DeZIM Graphics Style}
\usage{
dezim_style(
  title_size = 16,
  sub_title_size = 14,
  x_axis_title_size = 14,
  y_axis_title_size = 14,
  x_axis_text_size = 12,
  y_axis_text_size = 12
)

dezim_style(
  title_size = 16,
  sub_title_size = 14,
  x_axis_title_size = 14,
  y_axis_title_size = 14,
  x_axis_text_size = 12,
  y_axis_text_size = 12
)
}
\arguments{
\item{title_size}{Size of title}

\item{x_axis_title_size}{size of x axis title}

\item{y_axis_title_size}{size of y axis title}

\item{x_axis_text_size}{size of x axis text}

\item{y_axis_text_size}{size of y axis text}
}
\value{
No return, just a theme to add to ggplotfiles

No return, just a theme to add to ggplotfiles
}
\description{
DeZIM Graphics Style

DeZIM Graphics Style
}
