% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlechoice.R
\name{frequency_table}
\alias{frequency_table}
\title{Base table for single & multipe choice questions}
\usage{
frequency_table(data.table, group_by)
}
\arguments{
\item{data.table}{Output from either mutli or single summary}

\item{group_by}{Name of column in \code{dataset} used to partition the analysis into subgroups}
}
\value{
Gt table
}
\description{
Base table for single & multipe choice questions
}
