% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlechoice.R
\name{frequency_table}
\alias{frequency_table}
\title{Base table for single & multiple choice questions}
\usage{
frequency_table(data.table, group_by)
}
\arguments{
\item{data.table}{Output from either mutli or single summary}

\item{group_by}{Optional variable to group the analysis. If provided, the
frequencies and counts will be calculated within each subgroup.}
}
\value{
Gt table
}
\description{
Base table for single & multiple choice questions
}
