% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert.R
\name{likert_graph}
\alias{likert_graph}
\title{Quickly plot likert data (3,5,7,9 levels)}
\usage{
likert_graph(
  data,
  levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agre"),
  colors = c("#d73027", "#E36A64", "#FEEBD7", "#8ECF8C", "#66bd63")
)
}
\arguments{
\item{data}{Summerized likert data (first column must be names)}

\item{levels}{example text}

\item{colors}{example text}
}
\value{
example return
}
\description{
Quickly plot likert data (3,5,7,9 levels)
}
