% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert.R
\name{likert_graph}
\alias{likert_graph}
\title{Quickly plot likert data (3,5,7,9 levels)}
\usage{
likert_graph(
  dataset,
  question,
  labels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agre"),
  colors = c("#d73027", "#E36A64", "#FEEBD7", "#8ECF8C", "#66bd63", weights)
)
}
\arguments{
\item{dataset}{Dataframe or a tibble containing the \code{question} column to be analyzed}

\item{question}{The columns that contain each of the response options, can be selected by
using \strong{tidyselect} semanlatics or providing a vector of column names or numbers}
}
\value{
example return
}
\description{
Quickly plot likert data (3,5,7,9 levels)
}
