% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlechoice.R
\name{singlechoice_graph}
\alias{singlechoice_graph}
\title{Visualize frequencies in single choice survey questions}
\usage{
singlechoice_graph(
  dataset,
  question,
  subgroup = NULL,
  levels_to_exclude = NULL
)
}
\arguments{
\item{dataset}{data frame or a tibble}

\item{question}{name of the column in the dataset to graph}

\item{subgroup}{name of the subgroup to segment the data by}

\item{levels_to_exclude}{levels of the question varable to exclude in the graph}
}
\value{
A barchart of frequencies
}
\description{
Visualize frequencies in single choice survey questions
}
