% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlechoice.R
\name{singlechoice_table}
\alias{singlechoice_table}
\title{Summarize grouped counts and frequencies}
\usage{
singlechoice_table(
  dataset,
  question,
  group_by = NULL,
  subgroups_to_exclude = NULL,
  weights = NULL
)
}
\arguments{
\item{dataset}{Dataframe or a tibble containing survey question to be analyzed}

\item{question}{Name of the column in the \code{dataset} to be graphed}

\item{group_by}{Optional variable to stratify the frequencies of \code{question} variable}

\item{subgroups_to_exclude}{Optional vector of levels of \code{group_by} to exclude}

\item{weights}{Optional column containing survey weights}
}
\value{
table of frequences
}
\description{
Summarize grouped counts and frequencies
}
