% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{stringmatch}
\alias{stringmatch}
\title{Stringmatch}
\usage{
stringmatch(check_string, test_string)
}
\arguments{
\item{check_string}{placeholder}

\item{test_string}{placeholder}
}
\value{
placeholder
}
\description{
Stringmatch
}
